hardlockup: detect hard lockups without NMIs using secondary cpus
Emulate NMIs on systems where they are not available by using timer
interrupts on other cpus.  Each cpu will use its softlockup hrtimer
to check that the next cpu is processing hrtimer interrupts by
verifying that a counter is increasing.

This patch is useful on systems where the hardlockup detector is not
available due to a lack of NMIs, for example most ARM SoCs.
Without this patch any cpu stuck with interrupts disabled can
cause a hardware watchdog reset with no debugging information,
but with this patch the kernel can detect the lockup and panic,
which can result in useful debugging info.

Modified for older 2.6.39 kernel.

Changes since v1:
    renamed variables to clarify when referring to the next cpu
    factored out function to get next cpu
    fixed int vs unsigned int mismatches
    fixed race condition when offlining cpus

Change-Id: I0fc455beb1f249c3041e926d9e10da6a2b6feb9f
Signed-off-by: Colin Cross <ccross@android.com>
Signed-off-by: Takazumi Matsumoto <tdr@cyanogenmod.org>
Signed-off-by: flar2 <asegaert@gmail.com>